+++ src/parser.jl
@@ -3,13 +3,13 @@ using MacroTools
 """
     $(TYPEDSIGNATURES)
 
-Parse julia code into a [`FuzzyInferenceSystem`](@ref).
+Parse julia code into a [`MamdaniFuzzySystem`](@ref). See extended help for an example.
 
 # Extended help
 
 ### Example
 ```jldoctest
-fis = @fis function tipper(service, food)::tip
+fis = @mamfis function tipper(service, food)::tip
     service := begin
       domain = 0:10
       poor = GaussianMF(0.0, 1.5)
@@ -34,9 +34,9 @@ fis = @fis function tipper(service, food)::tip
     or = ProbSumOr
     implication = ProdImplication
 
-    service == poor || food == rancid => tip == cheap
-    service == good => tip == average
-    service == excellent || food == delicious => tip == generous
+    service == poor || food == rancid --> tip == cheap
+    service == good --> tip == average
+    service == excellent || food == delicious --> tip == generous
 
     aggregator = ProbSumAggregator
     defuzzifier = BisectorDefuzzifier
@@ -48,51 +48,128 @@ tipper
 
 Inputs:
 -------
-service ∈ [0, 10] with membership function
+service ∈ [0, 10] with membership functions:
     poor = GaussianMF{Float64}(0.0, 1.5)
     good = GaussianMF{Float64}(5.0, 1.5)
     excellent = GaussianMF{Float64}(10.0, 1.5)
 
-food ∈ [0, 10] with membership function
+food ∈ [0, 10] with membership functions:
     rancid = TrapezoidalMF{Int64}(-2, 0, 1, 3)
     delicious = TrapezoidalMF{Int64}(7, 9, 10, 12)
 
 
 Outputs:
 --------
-tip ∈ [0, 30] with membership function
+tip ∈ [0, 30] with membership functions:
     cheap = TriangularMF{Int64}(0, 5, 10)
     average = TriangularMF{Int64}(10, 15, 20)
     generous = TriangularMF{Int64}(20, 25, 30)
 
 
+Inference rules:
+----------------
+(service is poor ∨ food is rancid) --> tip is cheap
+service is good --> tip is average
+(service is excellent ∨ food is delicious) --> tip is generous
+
+
 Settings:
 ---------
-ProdAnd()
-
-ProbSumOr()
-
-ProdImplication()
-
-ProbSumAggregator()
-
-BisectorDefuzzifier(100)
+- ProdAnd()
+- ProbSumOr()
+- ProdImplication()
+- ProbSumAggregator()
+- BisectorDefuzzifier(100)
 ```
 """
-macro fis(ex::Expr)
-    return _fis(ex)
+macro mamfis(ex::Expr)
+    return _fis(ex, :MamdaniFuzzySystem)
 end
 
-const fis_settings = (:and, :or, :implication, :aggregator, :defuzzifier)
+"""
+    $(TYPEDSIGNATURES)
 
-function _fis(ex::Expr)
+Parse julia code into a [`SugenoFuzzySystem`](@ref). See extended help for an example.
+
+# Extended help
+
+### Example
+```jldoctest
+fis = @sugfis function tipper(service, food)::tip
+    service := begin
+        domain = 0:10
+        poor = GaussianMF(0.0, 1.5)
+        good = GaussianMF(5.0, 1.5)
+        excellent = GaussianMF(10.0, 1.5)
+    end
+
+    food := begin
+        domain = 0:10
+        rancid = TrapezoidalMF(-2, 0, 1, 3)
+        delicious = TrapezoidalMF(7, 9, 10, 12)
+    end
+
+    tip := begin
+        domain = 0:30
+        cheap = 0
+        average = food
+        generous = 2service, food, -2
+    end
+
+    service == poor && food == rancid --> tip == cheap
+    service == good --> tip == average
+    service == excellent || food == delicious --> tip == generous
+end
+
+# output
+
+tipper
+
+Inputs:
+-------
+service ∈ [0, 10] with membership functions:
+    poor = GaussianMF{Float64}(0.0, 1.5)
+    good = GaussianMF{Float64}(5.0, 1.5)
+    excellent = GaussianMF{Float64}(10.0, 1.5)
+
+food ∈ [0, 10] with membership functions:
+    rancid = TrapezoidalMF{Int64}(-2, 0, 1, 3)
+    delicious = TrapezoidalMF{Int64}(7, 9, 10, 12)
+
+
+Outputs:
+--------
+tip ∈ [0, 30] with membership functions:
+    cheap = 0
+    average = food
+    generous = 2service + food - 2
+
+
+Inference rules:
+----------------
+(service is poor ∧ food is rancid) --> tip is cheap
+service is good --> tip is average
+(service is excellent ∨ food is delicious) --> tip is generous
+
+
+Settings:
+---------
+- ProdAnd()
+- ProbSumOr()
+```
+"""
+macro sugfis(ex::Expr)
+    return _fis(ex, :SugenoFuzzySystem)
+end
+
+function _fis(ex::Expr, type)
     @capture ex function name_(argsin__)::({argsout__} | argsout__)
         body_
     end
-    inputs, outputs, opts, rules = parse_body(body, argsin, argsout)
+    inputs, outputs, opts, rules = parse_body(body, argsin, argsout, type)
 
-    fis = :(FuzzyInferenceSystem(; name = $(QuoteNode(name)), inputs = $inputs,
-                                 outputs = $outputs, rules = $rules))
+    fis = :($type(; name = $(QuoteNode(name)), inputs = $inputs,
+                  outputs = $outputs, rules = $rules))
     append!(fis.args[2].args, opts)
     return fis
 end
@@ -113,7 +190,7 @@ function parse_variable(var, args)
     return :($(QuoteNode(var)) => $ex)
 end
 
-function parse_body(body, argsin, argsout)
+function parse_body(body, argsin, argsout, type)
     opts = Expr[]
     rules = :(FuzzyRule[])
     inputs = :(dictionary([]))
@@ -124,20 +201,29 @@ function parse_body(body, argsin, argsout)
             if var in argsin
                 push!(inputs.args[2].args, parse_variable(var, args))
             elseif var in argsout
-                push!(outputs.args[2].args, parse_variable(var, args))
+                # TODO: makes this more scalable
+                push!(outputs.args[2].args,
+                      type == :SugenoFuzzySystem ? parse_sugeno_output(var, args, argsin) :
+                      parse_variable(var, args))
             else
                 throw(ArgumentError("Undefined variable $var"))
             end
         elseif @capture(line, var_=value_)
-            var in fis_settings ||
+            var in SETTINGS[type] ||
                 throw(ArgumentError("Invalid keyword $var in line $line"))
             push!(opts, Expr(:kw, var, value isa Symbol ? :($value()) : value))
         elseif @capture(line, ant_-->(cons__,) | cons__)
             push!(rules.args, parse_rule(ant, cons))
+        else
+            throw(ArgumentError("Invalid expression $line"))
         end
     end
     return inputs, outputs, opts, rules
 end
+
+#################
+# RULES PARSING #
+#################
 
 function parse_rule(ant, cons)
     Expr(:call, :FuzzyRule, parse_antecedent(ant), parse_consequents(cons))
@@ -161,4 +247,54 @@ function parse_consequents(cons)
         Expr(:call, :FuzzyRelation, QuoteNode(subj), QuoteNode(prop))
     end
     return Expr(:vect, newcons...)
+end
+
+############################
+# PARSE SUGENO EXPRESSIONS #
+############################
+
+function parse_sugeno_coeffs(exs::Vector, argsin::Vector, mfname::Symbol)
+    coeffs = :([$([0 for _ in 1:length(argsin)]...)])
+    offsets = []
+    for ex in exs
+        if @capture(ex, c_Number*var_Symbol)
+            idx = findfirst(==(var), argsin)
+            isnothing(idx) &&
+                throw(ArgumentError("Unkonwn variable $var in $mfname definition"))
+            coeffs.args[idx] = c
+        elseif ex isa Number
+            push!(offsets, ex)
+        elseif ex isa Symbol
+            idx = findfirst(==(ex), argsin)
+            isnothing(idx) &&
+                throw(ArgumentError("Unkonwn variable $ex in $mfname definition"))
+            coeffs.args[idx] = 1
+        else
+            throw(ArgumentError("Invalid expression $ex in $mfname definition"))
+        end
+    end
+    length(offsets) < 2 ||
+        throw(ArgumentError("multiple constants in $(Expr(:tuple, exs...))"))
+    coeffs, isempty(offsets) ? 0 : only(offsets)
+end
+
+function parse_sugeno_output(var, args, argsin)
+    mfs = :(dictionary([]))
+    ex = :(Variable())
+    inputs = Expr(:vect, map(QuoteNode, argsin)...)
+    for arg in args
+        if @capture(arg, domain=low_:high_)
+            push!(ex.args, :(Domain($low, $high)))
+        elseif @capture(arg, mfname_=c_Number)
+            push!(mfs.args[2].args, :($(QuoteNode(mfname)) => $(ConstantSugenoOutput(c))))
+        elseif @capture(arg, mfname_=(mfex__,) | mfex__)
+            coeffs, offset = parse_sugeno_coeffs(mfex, argsin, mfname)
+            mf = :(LinearSugenoOutput(Dictionary($inputs, $coeffs), $offset))
+            push!(mfs.args[2].args, :($(QuoteNode(mfname)) => $mf))
+        else
+            throw(ArgumentError("Invalid expression $arg"))
+        end
+    end
+    push!(ex.args, mfs)
+    return :($(QuoteNode(var)) => $ex)
 end